requirements.txt
torch
torchvision
pandas
dvc
notebook
numpy
matplotlib
seaborn
python-box
pyYAML
tqdm
ensure
joblib
types-PyYAML
scipy
Flask
Flask-Cors
gdown
-e .